+++ ble-server/config.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+user_path = Path(".")